+++ lib/shopify_cli/process_supervision.rb
@@ -50,19 +50,21 @@ module ShopifyCLI
       #
       # * `identifier` - a string or symbol to identify the new process by.
       # * `args` - a command to run, either a string or array of strings
+      # * `force_spawn` - whether we want the child process to be a spawn and not a fork, so it is terminated along with
+      #                   the parent
       #
       # #### Returns
       #
       # * `process` - ProcessSupervision instance if the process is running, this
       #   will be nil if the process did not start.
       #
-      def start(identifier, args)
+      def start(identifier, args, force_spawn: false)
         return for_ident(identifier) if running?(identifier)
 
         # Some systems don't support forking process without extra gems, so we resort to spawning a new child process -
         # that means that it dies along with the original process if it is interrupted. If possible, we fork the process
         # so that it doesn't have to be restarted on every run.
-        if Process.respond_to?(:fork)
+        if Process.respond_to?(:fork) && !force_spawn
           fork do
             pid_file = new(identifier, pid: Process.pid)
             pid_file.write